refactor(LineReader): Make use of the LineReader class in a much simpler way
The LineReader is made a concrete class whose declaration
is independent of the readline library.
Only its implementation depends on the presence or absence
of the readline library.
The LineReader code is embedded in a new library : line_reader,
to make the cli library independant of the USE_READLINE definition.
